refactor: top screen
